+++ Packages/Syntax/Sources/Syntax/Syntaxes/CSS/Queries/outline.scm
@@ -1,0 +1,14 @@
+;
+; outline.scm
+; for CSS
+;
+;  CotEditor
+;  https://coteditor.com
+;
+; © 2026 1024jp
+;
+
+; at-rules
+(at_rule) @outline.container
+(media_statement) @outline.container
+(keyframes_statement) @outline.container